pyproject.toml
[project]
name = "neural-style-transfer-latest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.4",
    "pillow>=11.3.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]
